requirements/prod.txt
PyYAML==5.3.1
deepdiff==5.0.2
easydict